requirements.txt
pdfplumber>=0.9.0
pandas>=1.5.0
python-dotenv>=0.21.0 